backend/requirements.txt
fastapi==0.115.0 
uvicorn==0.32.0 
python-dotenv==1.0.1 
firebase-admin==6.5.0 
pydantic==2.9.2 
python-jose[cryptography]==3.3.0 
passlib[bcrypt]==1.7.4 
python-multipart==0.0.12 
google-cloud-storage==2.18.2 
aiohttp==3.10.10 
pytest==8.3.3 
httpx==0.27.2